+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Includes latest features and improvements from production in the Firestore Emulator.
 - Fixes issue where all database functions triggered on the default namespace (#2501)
 - Fixes issue where rules paths were not normalized before reading (#2544)
+- Functions emulator waits for all functions to finish before exiting (#1813)